requirements.txt
selenium
webdriver-manager
psycopg2-binary
streamlit
pandas